23.08.24

+++ app/src/main/java/com/decide/app/feature/assay/assayProcess/ui/Answers.kt
@@ -3,5 +3,5 @@ package com.decide.app.feature.assay.assayProcess.ui
 data class Answers(
     val idQuestion: Int,
     val idAnswer: Int,
-    val answerValue: Int
+    val answerValue: Float
 )
